test-requirements.txt
testcontainers[mysql]
cryptography
pytest
mysql-connector-python
